# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel APIs",
        "description": "Sample APIs for praticepractice mock. | Postman - https://documenter.getpostman.com/view/2666389/2sA2r6YQVS",
        "contact": {
            "email": "yhvreddydev@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/user": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get all Users List",
                "operationId": "babaf8a5ff0bc9f2e918c6fe96b2ade8",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Creating new user",
                "operationId": "b06edc54e183ec3754acab6da6f77d26",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "User name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Failed to create user"
                    }
                }
            }
        },
        "/api/v1/user/{user}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get logged-in user details",
                "operationId": "eaf690f9636aec5184475b8f76859e82",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update existing user",
                "operationId": "af28c6730f6cf535c0bebbda77da348e",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User Id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "User name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User updated successful"
                    },
                    "401": {
                        "description": "Failed to update user"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Hard delete user details",
                "operationId": "db5180547ea70143a55d55a175f97799",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/v1/user/{user}/trashed": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Soft delete user details",
                "operationId": "21001f13d46ebc91b23b6d4b85cf1fec",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/v1/user/{user}/restore": {
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Restore soft deleted user details",
                "operationId": "84fa4b3b1d8bafa3c96f1cb958b0625c",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Authenticate user and generate token",
                "operationId": "c176298c4d97aeca15babe5cc6399df1",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/v1/todo": {
            "get": {
                "tags": [
                    "Todo"
                ],
                "summary": "Get all Todo's List",
                "operationId": "cef9decc230a3cecc9f611819ba49c7f",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Todo"
                ],
                "summary": "Creating new todo task",
                "operationId": "9daa1a8cd72e60e673aa038794b5b613",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Todo title",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Todo description",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "User Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todo added successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Failed to add todo"
                    }
                }
            }
        },
        "/api/v1/todo/{todo}": {
            "put": {
                "tags": [
                    "Todo"
                ],
                "summary": "Update todo task",
                "operationId": "220af1cc0af8b3c8d03f2a4688fa746c",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Todo title",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Todo description",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "todo",
                        "in": "path",
                        "description": "Todo Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todo Update successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Failed to update todo"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Todo"
                ],
                "summary": "Hard delete todo details",
                "operationId": "dd7e08dab5e1b7bce902957d1ad6ed0b",
                "parameters": [
                    {
                        "name": "todo",
                        "in": "path",
                        "description": "Todo id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/v1/todo/{todo}/trashed": {
            "delete": {
                "tags": [
                    "Todo"
                ],
                "summary": "Soft delete todo details",
                "operationId": "3852e57f76e316f17429d5df3ace655d",
                "parameters": [
                    {
                        "name": "todo",
                        "in": "path",
                        "description": "Todo id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/v1/todo/{todo}/restore": {
            "patch": {
                "tags": [
                    "Todo"
                ],
                "summary": "Restore soft deleted todo details",
                "operationId": "f028135c7ac570f508da6a3ae9a6f00e",
                "parameters": [
                    {
                        "name": "todo",
                        "in": "path",
                        "description": "Todo id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            }
        }
    }
}